basic new landing page (#71)
