+++ components/standard/SignInButton.tsx
@@ -1,16 +1,16 @@
 import {signIn} from "next-auth/react";
 import {FaGoogle} from "react-icons/fa";
-import React from "react";
+import React, {ReactNode} from "react";
 import UiButton from "../style/UiButton";
 
-export default function SignInButton() {
+export default function SignInButton({className, children}: {className?: string, children?: ReactNode}) {
     return (
         <UiButton
-            className="up-button primary"
+            className={`up-button primary ${className || ""}`}
             onClick={() => signIn("google")}
         >
             <div className="flex items-center">
-                <FaGoogle/><span className="ml-2">Sign in</span>
+                <FaGoogle/><span className="ml-2">{children || "Sign in"}</span>
             </div>
         </UiButton>
     );
